requirements.txt
google-generativeai
requests
Flask
python-dotenv